Use void * for functions that access bytes of object representation.
Although out() and secaddbytes() only use their argument to call
fwrite()/memcpy(), using the type uint8_t * suggests that it might
be dereferenced as uint8_t. In general, it is undefined behavior
to access an object through a uint8_t pointer, since uint8_t is not
necessarily a character type.

So, just use void * to avoid some casts and implication of possible
undefined behavior.