Fix cut-off check for 1-byte jump offset
The size of a 1-byte jump is 2, so the offset to the end of a the
jump for a negative distance is `distance - 2`. Similarly, the size
of a 4-byte jump is 5 or 6, so the offset from the end of the 4-byte
jump is `distance - 5` or `distance - 6`.

Also, add an assert to make sure the offset really does fit in 1 byte.